Created a crossplatform test project.
